tollgate/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tollgate"
version = "1.4.0"
description = "Runtime policy enforcement engine for financial-services AI agents — ALLOW / ASK / DENY with a full audit trail"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Ravi Labs", email = "ravi@ravi-labs.dev" },
]
keywords = [
    "ai", "agents", "policy", "governance", "compliance",
    "financial-services", "erisa", "llm", "tollgate"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: AsyncIO",
]
dependencies = [
    "pyyaml>=6.0.0",
    "click>=8.0.0",
]

[project.scripts]
tollgate = "tollgate.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# ── Storage backends ──────────────────────────────────────────────────────────
redis = [
    "redis[hiredis]>=5.0.0",
]
aws = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
]

# ── Observability ─────────────────────────────────────────────────────────────
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

# ── Security / encryption ─────────────────────────────────────────────────────
encryption = [
    "cryptography>=41.0.0",
]

# ── Framework integrations ────────────────────────────────────────────────────
mcp = [
    "mcp>=0.1.0",
]
strands = [
    "strands-agents>=0.1.0",
    "tollgate[aws]",
]

# ── Bundles ───────────────────────────────────────────────────────────────────
full = [
    "tollgate[redis,aws,otel,encryption]",
]

[project.urls]
Homepage = "https://github.com/ravi-labs/agent-incubation/tree/main/tollgate"
Repository = "https://github.com/ravi-labs/agent-incubation.git"
Documentation = "https://github.com/ravi-labs/agent-incubation/tree/main/tollgate#readme"
Issues = "https://github.com/ravi-labs/agent-incubation/issues"
"Agent Foundry" = "https://github.com/ravi-labs/agent-incubation/tree/main/agent-foundry"

[tool.hatch.build.targets.wheel]
packages = ["src/tollgate"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG", "PTH", "RET"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG001", "ARG002", "E402", "E501", "B007"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
